# service/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v2/reserve/add:
        post:
            tags:
                - TimeTableService
            operationId: TimeTableService_AddTimeTables
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.Ordered'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.Ordered'
    /v2/reserve/delete/{id}:
        delete:
            tags:
                - TimeTableService
            operationId: TimeTableService_DeleteTimeTable
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.DeleteTimeTableResponse'
    /v2/reserve/get_ordered/{orderedId}:
        get:
            tags:
                - TimeTableService
            operationId: TimeTableService_GetOrderedByOrderedId
            parameters:
                - name: orderedId
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.Ordered'
    /v2/reserve/get_time_tables/{userId}/{day}:
        get:
            tags:
                - TimeTableService
            operationId: TimeTableService_GetTimeTablesOrdered
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
                - name: day
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.Ordered'
    /v2/reserve/update:
        patch:
            tags:
                - TimeTableService
            operationId: TimeTableService_UpdateTimeTable
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.TimeTable'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.TimeTable'
    /v2/times/add:
        post:
            tags:
                - TimeService
            operationId: TimeService_CreateTime
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.TimeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.Times'
    /v2/times/list{id}:
        get:
            tags:
                - TimeService
            operationId: TimeService_ListTimes
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.TimeList'
    /v2/times/update:
        patch:
            tags:
                - TimeService
            operationId: TimeService_UpdateTime
            parameters:
                - name: id
                  in: query
                  schema:
                    type: string
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: timeSpace
                  in: query
                  schema:
                    type: string
                - name: startTime
                  in: query
                  schema:
                    type: string
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.Times'
    /v2/times/{id}:
        delete:
            tags:
                - TimeService
            operationId: TimeService_DeleteTime
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.DeleteTimeResponse'
components:
    schemas:
        blackpig.v2.DeleteTimeResponse:
            type: object
            properties:
                success:
                    type: boolean
        blackpig.v2.DeleteTimeTableResponse:
            type: object
            properties:
                success:
                    type: boolean
...
        blackpig.v2.Ordered:
            type: object
            properties:
                timeTables:
                    type: array
                    items:
                        $ref: '#/components/schemas/blackpig.v2.TimeTable'
...
        blackpig.v2.TimeList:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: integer
                    format: int32
                times:
                    type: array
                    items:
                        $ref: '#/components/schemas/blackpig.v2.Times'
        blackpig.v2.TimeRequest:
            type: object
            properties:
                userId:
                    type: integer
                    format: int32
                timeSpace:
                    type: string
                startTime:
                    type: string
        blackpig.v2.TimeTable:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                startTime:
                    type: string
                endTime:
                    type: string
                day:
                    type: string
                userId:
                    type: integer
                    format: int32
                orderedId:
                    type: integer
                    format: int32
                createdAt:
                    type: string
                updatedAt:
                    type: string
        blackpig.v2.Times:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: integer
                    format: int32
                timeSpace:
                    type: string
                startTime:
                    type: string
...